agent/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
fastapi==0.128.0
h11==0.16.0
httptools==0.7.1
idna==3.11
groq==0.32.0
edge-tts==6.1.12
langchain==0.3.0
langchain-openai==0.2.0
langchain-core==0.3.0
langgraph>=0.2.0
python-dotenv==1.2.1
PyYAML==6.0.3
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
